% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensval.R
\docType{data}
\name{bangladesh_floods}
\alias{bangladesh_floods}
\title{This is data to be included in my package.}
\format{
An object of class \code{list} of length 3.
}
\usage{
bangladesh_floods
}
\description{
This is data to be included in my package.
}
\references{
del Ninno et al. (2001), \url{https://doi.org/10.7910/DVN/UJA9N5}
}
\keyword{datasets}
