% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensval.R
\name{is_binary}
\alias{is_binary}
\title{Classifies numeric variables as binary or continuous.}
\usage{
is_binary(x)
}
\arguments{
\item{x}{pair-matched or individual-level data.}
}
\value{
Boolean indicator of whether the variable is binary.
}
\description{
This function is internal to other functions and checks if a particular
numeric variable should be treated as binary or continuous.
}
