% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensval.R
\name{sensitivity_value}
\alias{sensitivity_value}
\title{Computes sensitivity value for pair-matched data.}
\usage{
sensitivity_value(mpdifs, a = a, alt = alt)
}
\arguments{
\item{mpdifs}{vector of length nsets, where entry k contains the treated
minus control value corresponding to the kth matched pair.}

\item{a}{alpha-level of test.}

\item{alt}{character denoting direction of alternative hypothesis.}
}
\value{
Sensitivity value.
}
\description{
This function is internal to other functions and calculates the sensitivity
value of pair-matched data using numerical methods.
}
