% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensval.R
\name{sv}
\alias{sv}
\title{Sens-Val method for pair-matched data.}
\usage{
sv(
  data_outcomes,
  control_idx,
  treated_idx,
  alt_vec = NULL,
  planning_sample_prop = 0.2,
  alpha = 0.05,
  gamma = 1,
  nboot = 250,
  seed = 0
)
}
\arguments{
\item{data_outcomes}{dataframe whose rows contain N individual subjects and
whose columns have L particular hypotheses.}

\item{control_idx}{matrix of size nsetsx1, where entry k contains the control
index of data_outcomes corresponding to the kth matched pair.}

\item{treated_idx}{vector of length nsets, where entry k contains the treated
index of data_outcomes corresponding to the kth matched pair.}

\item{alt_vec}{character vector of size L denoting direction of one-sided
alternative hypotheses.}

\item{planning_sample_prop}{proportion of the data to allocate to the
planning sample.}

\item{alpha}{alpha-level of test.}

\item{gamma}{value of the sensitivity parameter.}

\item{nboot}{number of bootstraps to calculate variance of each sensitivity
value on the planning sample.}

\item{seed}{value of seed for reproducibility.}
}
\value{
A list containing the names of hypotheses the Sens-Val procedure
chooses to test and those it rejects.
}
\description{
This function runs the Sens-Val procedure, introduced by Bekerman et al.
(2024), on pair-matched data from an observational study. The method
identifies and tests hypotheses from a split sample that are robust to
potential unmeasured biases. As defaults, binary data are analyzed using
McNemar's test statistic and continuous data are analyzed using Wilcoxon's
signed rank statistic.
}
